+++ playground/src/main.js
@@ -4,9 +4,4 @@ const app = new App({
   target: document.getElementById("app")
 });
 
-const bootScreen = document.getElementById("boot-screen");
-if (bootScreen) {
-  bootScreen.remove();
-}
-
 export default app;